+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/transformers/ContributionHintIrTransformer.kt
@@ -86,8 +86,9 @@ internal class ContributionHintIrTransformer(
 
       val fileNameWithoutExtension =
         sequence {
-            yieldAll(Symbols.FqNames.metroHintsPackage.pathSegments())
-            yieldAll(declaration.classIdOrFail.relativeClassName.pathSegments())
+            val classId = declaration.classIdOrFail
+            yieldAll(classId.packageFqName.pathSegments())
+            yield(classId.joinSimpleNames(separator = "").shortClassName)
             yield(callableName)
           }
           .joinToString(separator = "") { it.asString().capitalizeUS() }